demos/full-example/requirements.txt
# General
apache-airflow>=2.10.0,<3
click>=8.0
jinja2>=3.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyspark>=3.5.0
boto3>=1.34.0
python-dotenv>=1.0.0

# Data Quality
soda-core>=4.0.0
soda-postgres>=4.0.0

# Tests
pytest>=7.4.0